# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI-powered E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterBody'
      responses:
        201:
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginBody'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
                quantity:
                  type: integer
              required: [sessionId, quantity]
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /negotiation/start:
    post:
      operationId: startNegotiation
      tags: [negotiation]
      summary: Start a price negotiation session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartNegotiationBody"
      responses:
        "200":
          description: Negotiation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NegotiationSession"
  /negotiation/offer:
    post:
      operationId: makeOffer
      tags: [negotiation]
      summary: Make a counter-offer in negotiation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MakeOfferBody"
      responses:
        "200":
          description: Counter-offer response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NegotiationResponse"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        category:
          type: string
        brand:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        specs:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
        demandScore:
          type: number
        createdAt:
          type: string
      required: [id, name, price, category, stock, rating, reviewCount]
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        negotiatedPrice:
          type: number
      required: [id, productId, quantity]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required: [items, total, itemCount]
    AddToCartBody:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
        negotiatedPrice:
          type: number
      required: [sessionId, productId, quantity]
...
    StartNegotiationBody:
      type: object
      properties:
        productId:
          type: integer
        sessionId:
          type: string
        offerPrice:
          type: number
      required: [productId, sessionId, offerPrice]
    NegotiationSession:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        originalPrice:
          type: number
        currentOffer:
          type: number
        status:
          type: string
          enum: [active, accepted, rejected, countered]
        message:
          type: string
        roundsLeft:
          type: integer
      required: [sessionId, productId, originalPrice, currentOffer, status, message, roundsLeft]
    MakeOfferBody:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        offerPrice:
          type: number
        previousOffer:
          type: number
      required: [sessionId, productId, offerPrice]
    NegotiationResponse:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected, countered]
        counterOffer:
          type: number
        finalPrice:
          type: number
        message:
          type: string
        roundsLeft:
          type: integer
      required: [status, message, roundsLeft]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
      required: [id, name, email, role]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
      required: [user, token]